# docs/PLUGIN_API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: Renderiq Plugin API
  description: |
    Unified API for Renderiq plugins across all AEC platforms (SketchUp, Revit, AutoCAD, Rhino, ArchiCAD, Blender).
    
    ## Authentication
    - Bearer token: Use `Authorization: Bearer <token>` header
    - API key: Use `X-Api-Key: <key>` header (alternative to Bearer token)
    
    ## Webhooks
    Register webhooks to receive render status updates without polling. Webhook payloads are signed with HMAC-SHA256.
    
    ## Resumable Uploads
    For large files (>50MB), use resumable upload endpoints to upload in chunks.
    
  version: "1.0.0"
  contact:
    name: Renderiq Support
    email: support@renderiq.io
    url: https://renderiq.io/support
servers:
  - url: https://renderiq.io/api/plugins
    description: Production
  - url: http://localhost:3000/api/plugins
    description: Development (local)
security:
  - bearerAuth: []
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from `/auth/signin`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key for plugin authentication (alternative to Bearer token)
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid authentication token"
        errorCode:
          type: string
          example: "AUTH_REQUIRED"
        details:
          type: object
          description: Additional error details
...
    CreateRenderRequest:
      type: object
      required:
        - uploadedImageData
        - uploadedImageType
      properties:
        prompt:
          type: string
          description: Optional prompt (system prompt generated if not provided)
          maxLength: 2000
        uploadedImageData:
          type: string
          format: binary
          description: Image data (base64 encoded or binary)
        uploadedImageType:
          type: string
          enum: [image/png, image/jpeg, image/jpg]
          example: "image/png"
        quality:
          type: string
          enum: [standard, high, ultra]
          default: high
          description: |
            - standard: 1080p (5 credits)
            - high: 4K (10 credits)
            - ultra: 4K Enhanced (15 credits)
        aspectRatio:
          type: string
          enum: [16:9, 4:3, 1:1, 9:16, 3:2, 2:3]
          default: "16:9"
        style:
          type: string
          enum: [photorealistic, dramatic, soft, studio, natural]
          default: photorealistic
        model:
          type: string
          enum: [gemini-3-pro-image-preview, gemini-2.5-flash-image]
          description: AI model to use
        projectId:
          type: string
          format: uuid
          description: Optional project ID to associate render with
        callback_url:
          type: string
          format: uri
          description: Webhook URL to receive render status updates
        resumable:
          type: boolean
          default: false
          description: Use resumable upload for large files (>50MB)
    CreateRenderResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            renderId:
              type: string
              format: uuid
            status:
              type: string
              enum: [pending, processing]
              example: "pending"
            creditsCost:
              type: integer
              example: 10
            estimatedTime:
              type: integer
              description: Estimated processing time in seconds
              example: 30
            uploadUrl:
              type: string
              format: uri
              description: Resumable upload URL (if resumable=true)
...
paths:
...
  /renders:
    post:
      tags:
        - Renders
      summary: Create a new render request
      description: |
        Submit an image for AI rendering. Supports both standard and resumable uploads.
        
        **Resumable Uploads**: For files >50MB, set `resumable=true` to get an upload URL for chunked uploads.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateRenderRequest'
      responses:
        "202":
          description: Render request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRenderResponse'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  errorCode:
                    type: string
                    example: "INSUFFICIENT_CREDITS"
                  requiredCredits:
                    type: integer
                  availableCredits:
                    type: integer
...